+++ benight-admin/src/app/navigator/detalle/detalle-info-event/detalle-info-event-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core'
 import { Routes, RouterModule } from '@angular/router'
-import { DetalleInfoPage } from './detalle-info.page'
+import { DetalleInfoEventPage } from './detalle-info-event.page'
 
 const routes: Routes = [
   {
     path: '',
-    component: DetalleInfoPage
+    component: DetalleInfoEventPage
   }
 ]
 
@@ -13,4 +13,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class DetalleInfoRoutingModule { }
+export class DetalleInfoEventRoutingModule { }