+++ benight-admin/src/app/navigator/detalle/detalle-info-event/detalle-info-event.page.spec.ts
@@ -1,22 +1,22 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core'
 import { async, ComponentFixture, TestBed } from '@angular/core/testing'
 
-import { DetalleInfoPage } from './detalle-info.page'
+import { DetalleInfoEventPage } from './detalle-info-event.page'
 
-describe('DetalleInfoPage', () => {
-  let component: DetalleInfoPage
-  let fixture: ComponentFixture<DetalleInfoPage>
+describe('DetalleInfoEventPage', () => {
+  let component: DetalleInfoEventPage
+  let fixture: ComponentFixture<DetalleInfoEventPage>
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ DetalleInfoPage ],
+      declarations: [ DetalleInfoEventPage ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
     })
     .compileComponents()
   }))
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(DetalleInfoPage)
+    fixture = TestBed.createComponent(DetalleInfoEventPage)
     component = fixture.componentInstance
     fixture.detectChanges()
   })