+++ benight-admin/src/app/navigator/gestion/eventos/eventos.page.ts
@@ -1,5 +1,10 @@
 import { Component, OnInit } from '@angular/core'
 import { SharedDataService } from '@bn8-services/shared-data.service'
+import { DbService } from '@bn8-services/db.service'
+import { AuthService } from '@bn8-services/auth.service'
+import { Observable } from 'rxjs'
+
+
 
 @Component({
   selector: 'gestion-eventos',
@@ -8,10 +13,25 @@ import { SharedDataService } from '@bn8-services/shared-data.service'
 })
 export class EventosPage implements OnInit {
   
-  constructor(private sd: SharedDataService) {
+  events: Observable<any>
+
+  constructor(
+    private sd: SharedDataService,
+    private db:DbService,
+    private auth:AuthService,
+
+    ) {}
+
+  async ngOnInit() {
+    this.sd.set('header', 'Eventos')
+    //let user = await this.auth.user$
+    let user = 0
+    this.events = await this.db.leftJoin('propietario_eventos','eventos','ownerUid', user, 'event','event')
+    console.log(this.events)
   }
 
-  ngOnInit() {
-    this.sd.set('header', 'Eventos')
+  trackById(idx, todo) {
+    return todo.id
   }
+
 }